+++ server/src/admin/admin.service.ts
@@ -5,21 +5,20 @@ import {
 } from '@nestjs/common';
 import type { Response } from 'express';
 import type { RegisterAdminDto } from './dto/register-admin.dto';
-import { AccountRepository } from './account.repository';
+import { AdminRepository } from './admin.repository';
 import * as bcrypt from 'bcrypt';
 import { cookieConfig } from '../common/cookie/cookie.config';
 import * as uuid from 'uuid';
 import { RedisService } from '../redis/redis.service';
 import type { LoginAdminDto } from './dto/login-admin.dto';
 
-
 @Injectable()
-export class AccountService {
+export class AdminService {
   // 12시간 후 자동 만료
   private readonly SESSION_TTL = 60 * 60 * 12;
 
   constructor(
-    private readonly loginRepository: AccountRepository,
+    private readonly loginRepository: AdminRepository,
     private readonly redisService: RedisService,
   ) {}
