+++ server/src/app.module.ts
@@ -3,7 +3,7 @@ import { winstonModule } from './common/logger/logger.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { loadDBSetting } from './common/database/load.config';
-import { AccountModule } from './account/account.module';
+import { AdminModule } from './admin/admin.module';
 import { RedisModule } from './redis/redis.module';
 import { RssModule } from './rss/rss.module';
 import { FeedModule } from './feed/feed.module';
@@ -23,7 +23,7 @@ import { FeedModule } from './feed/feed.module';
       useFactory: (configService: ConfigService) =>
         loadDBSetting(configService),
     }),
-    AccountModule,
+    AdminModule,
     RedisModule,
     RssModule,
     FeedModule,
